requirements.txt
neat-python>=0.92
numpy>=1.21.0
scikit-learn>=1.0.0
plotly>=5.3.0
networkx>=2.6.0
folium>=0.12.0
umap-learn>=0.5.0
pandas>=1.3.0
matplotlib>=3.4.0
kaleido>=0.2.1  # Para exportar plotly como PNG/SVG
